Fix `class-object/c` for classes that implement interfaces
Instantiation of objects for classes that implement interfaces
clobber the `make-object` field of class. If we've proxied it,
that gets blown away. The solution is to pretend like the class
doesn't have interface contracts and then in the proxied
`make-object` do what was supposed to happen in the appropriate
branch of `fetch-concrete-class` (along with the object contract
check).